+++ libswirl/deps/libzip/zip_source_filep.c
@@ -115,7 +115,7 @@ read_file(void *state, void *data, size_t len, enum zip_source_cmd cmd)
 {
     struct read_file *z;
     char *buf;
-    int i, n;
+    size_t i, n;
 
     z = (struct read_file *)state;
     buf = (char *)data;
